refactor: Refactor DTO, Request, and Response Structure for Better Separation of Concerns
