+++ backend/chidic-common/src/main/kotlin/com/chidicapp/api/request/CommentRequest.kt
@@ -1,4 +1,4 @@
-package com.chidiccommon.dto
+package com.chidicapp.api.request
 
 data class CommentRequest(
     val content: String
