+++ backend/chidic-domain/src/main/kotlin/com/chidicdomain/dto/CommentCreateDto.kt
@@ -1,0 +1,7 @@
+package com.chidicdomain.dto
+
+data class CommentCreateDto (
+    val userId: Long,
+    val feedPostId: Long,
+    val content: String
+)
